requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
torchaudio>=2.0.0
accelerate>=0.20.0
sentencepiece>=0.1.97

# Extended model support
tokenizers>=0.13.0
safetensors>=0.3.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
h5py>=3.7.0

# Scientific computing
scipy>=1.9.0

# Natural language processing
spacy>=3.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Dimensionality reduction and ML
scikit-learn>=1.1.0
umap-learn>=0.5.0

# Web interface
flask>=2.3.0
werkzeug>=2.3.0

# System monitoring
psutil>=5.9.0

# Configuration
PyYAML>=6.0

# Progress bars
tqdm>=4.64.0

# Ollama client
ollama>=0.1.0

# System monitoring
psutil>=5.9.0
requests>=2.28.0

# Optional GPU monitoring
pynvml>=11.4.0

# API and web dependencies (optional)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
aiohttp>=3.8.0
websockets>=11.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
flask>=2.3.0
jinja2>=3.1.0

# Advanced analysis dependencies
networkx>=3.0
igraph>=0.10.0
community>=1.0.0

# Neural network interpretability
captum>=0.6.0
gradio>=3.0.0

# Additional utilities
click>=8.0.0
rich>=13.0.0
typer>=0.9.0
jsonlines>=3.1.0

# Statistical analysis
statsmodels>=0.14.0
pingouin>=0.5.0

# Causal analysis  
dowhy>=0.10.0
econml>=0.14.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
hypothesis>=6.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0
